Gather various overclocking logics into a single class
