+++ src/main/java/gregtech/api/capability/impl/RecipeLogicSteam.java
@@ -28,6 +28,8 @@ import net.minecraftforge.fluids.IFluidTank;
 
 import javax.annotation.Nonnull;
 
+import static gregtech.api.recipes.logic.OverclockingLogic.standardOverclockingLogic;
+
 public class RecipeLogicSteam extends AbstractRecipeLogic implements IVentable {
 
     private final IFluidTank steamFluidTank;
